+++ bw.core/src/main/java/sa/gov/nic/bio/bw/core/utils/GuiUtils.java
@@ -218,8 +218,12 @@ public class GuiUtils implements AppLogger
 					
 					try
 					{
-						BclUtils.launchAppByBCL(serverUrl, AppConstants.APP_CODE.toLowerCase(), -1,
-						                        -1, null);
+						String jvmArch = System.getProperty("sun.arch.data.model");
+						boolean jvmArch32 = "32".equals(jvmArch); // consider everything else as 64-bit
+						String appCode = AppConstants.APP_CODE.toLowerCase();
+						if(!jvmArch32) appCode = appCode + "64";
+						
+						BclUtils.launchAppByBCL(serverUrl, appCode, -1, -1, null);
 					}
 					catch(Exception e)
 					{
@@ -426,8 +430,14 @@ public class GuiUtils implements AppLogger
 		comboBox.showingProperty().addListener(showingPropertyChangeListener);
 		comboBox.getEditor().textProperty().addListener(textPropertyChangeListener);
 	}
-
+	
 	public static void attachImageDialog(CoreFxController coreFxController, ImageView imageView, String dialogTitle,
+	                                     String showImageText, boolean blur)
+	{
+		attachImageDialog(coreFxController.getStage(), imageView, dialogTitle, showImageText, blur);
+	}
+
+	public static void attachImageDialog(Stage stage, ImageView imageView, String dialogTitle,
 	                                     String showImageText, boolean blur)
 	{
 		Runnable runnable = () ->
@@ -462,14 +472,14 @@ public class GuiUtils implements AppLogger
 
 			BorderPane borderPane = new BorderPane();
 			borderPane.setCenter(iv);
-			Stage stage = DialogUtils.buildCustomDialog(coreFxController.getStage(), dialogTitle, borderPane,
+			Stage newStage = DialogUtils.buildCustomDialog(stage, dialogTitle, borderPane,
 			                                            Context.getGuiLanguage().getNodeOrientation()
 					                                            == NodeOrientation.RIGHT_TO_LEFT, false);
-			stage.getScene().addEventHandler(KeyEvent.KEY_PRESSED, keyEvent ->
+			newStage.getScene().addEventHandler(KeyEvent.KEY_PRESSED, keyEvent ->
 			{
-				if(keyEvent.getCode() == KeyCode.ESCAPE) stage.close();
+				if(keyEvent.getCode() == KeyCode.ESCAPE) newStage.close();
 			});
-			stage.show();
+			newStage.show();
 		};
 
 		imageView.setOnMouseClicked(mouseEvent ->