+++ RELEASE-NOTES.md
@@ -11,6 +11,16 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 
 ---
 
+## v2020.03.3
+
+- NEW: Latent Reverse Search Workflow.
+- ENH: Upgrade Java SE to 14.
+- ENH: Upgrade Java FX to 14.
+- ENH: Upgrade dependencies to latest versions.
+- NEW: Add support for web-start 64-bit deployment.
+
+---
+
 ## v2020.03.2
 
 - NEW: Add new NIC logo.